hackathon_website/backend/requirements.txt
fastapi==0.104.1
finnhub-python==2.4.20
uvicorn==0.24.0
httpx==0.27.2
aiohttp==3.9.1
python-dotenv==1.0.1
pydantic==2.5.0